+++ google/api_versions_test.go
@@ -3,14 +3,20 @@ package google
 import "testing"
 
 type ExpectedApiVersions struct {
-	Create     ComputeApiVersion
-	ReadDelete ComputeApiVersion
-	Update     ComputeApiVersion
+	Create     ApiVersion
+	ReadDelete ApiVersion
+	Update     ApiVersion
 }
 
-func TestComputeApiVersion(t *testing.T) {
+func TestApiVersion(t *testing.T) {
 	baseVersion := v1
 	betaVersion := v0beta
+	maxTestApiVersion := func(versionsInUse map[ApiVersion]struct{}) ApiVersion {
+		if _, ok := versionsInUse[betaVersion]; ok {
+			return betaVersion
+		}
+		return baseVersion
+	}
 
 	cases := map[string]struct {
 		Features         []Feature
@@ -173,7 +179,7 @@ func TestComputeApiVersion(t *testing.T) {
 			FieldsWithHasChange: keys,
 		}
 
-		apiVersion := getComputeApiVersion(d, v1, tc.Features)
+		apiVersion := getApiVersion(d, v1, tc.Features, maxTestApiVersion)
 		if apiVersion != tc.ExpectedApiVersions.Create {
 			t.Errorf("bad: %s, Expected to see version %v for create, got version %v", tn, tc.ExpectedApiVersions.Create, apiVersion)
 		}
@@ -184,7 +190,7 @@ func TestComputeApiVersion(t *testing.T) {
 			FieldsInSchema: tc.FieldsInSchema,
 		}
 
-		apiVersion = getComputeApiVersion(d, v1, tc.Features)
+		apiVersion = getApiVersion(d, v1, tc.Features, maxTestApiVersion)
 		if apiVersion != tc.ExpectedApiVersions.ReadDelete {
 			t.Errorf("bad: %s, Expected to see version %v for read/delete, got version %v", tn, tc.ExpectedApiVersions.ReadDelete, apiVersion)
 		}
@@ -196,7 +202,7 @@ func TestComputeApiVersion(t *testing.T) {
 			FieldsWithHasChange: tc.UpdatedFields,
 		}
 
-		apiVersion = getComputeApiVersionUpdate(d, v1, tc.Features, tc.UpdateOnlyFields)
+		apiVersion = getApiVersionUpdate(d, v1, tc.Features, tc.UpdateOnlyFields, maxTestApiVersion)
 		if apiVersion != tc.ExpectedApiVersions.Update {
 			t.Errorf("bad: %s, Expected to see version %v for update, got version %v", tn, tc.ExpectedApiVersions.Update, apiVersion)
 		}
@@ -206,6 +212,7 @@ func TestComputeApiVersion(t *testing.T) {
 type ResourceDataMock struct {
 	FieldsInSchema      map[string]interface{}
 	FieldsWithHasChange []string
+	id                  string
 }
 
 func (d *ResourceDataMock) HasChange(key string) bool {
@@ -227,4 +234,17 @@ func (d *ResourceDataMock) GetOk(key string) (interface{}, bool) {
 	}
 
 	return nil, false
+}
+
+func (d *ResourceDataMock) Set(key string, value interface{}) error {
+	d.FieldsInSchema[key] = value
+	return nil
+}
+
+func (d *ResourceDataMock) SetId(v string) {
+	d.id = v
+}
+
+func (d *ResourceDataMock) Id() string {
+	return d.id
 }